Fix typos in azure-examples.md
Fix two small typos in the azure-examples.md.

Also, I wanted to note that the correct capitalization of Azure DevOps includes capitalizing the 'O' in DevOps.  Throughout this page and other pages in the Snyk documentation, it is spelled with a lowercase 'o' as 'Azure Devops.'  I did not change this page to include a capital 'O' as you have other pages with the same lowercase 'o' in the documentation.  This is also a minor correction to the name and may not be worth the effort to update.