chore: document cli auth defaults in upcoming release
Explain that the CLI uses OAuth by default (after upcoming stable
release).

Document how to fall back on interactive token auth if necessary, and
that it's deprecated and will be removed.

Targeting 1.1293 for the next release that will implement these changes.

Consequently this PR should be held back until 1.1293.x-stable is released.

CLI-458