docs: fix V1 update existing integration docs
Some of the integrations endpoints are polymorphic, where what you send as the body, changes how the endpoint works.
The examples were not being reflected in the reference docs to show what configurations could be used, so I've changed
it to all the parameters, and we will let the docs explain how to configure the call to perform a given task, e.g.
deleting a brokered integration.