Update clone URL in gemini-cli-guide.md to HTTPS
Using an https:// rather than git:// URL means that SSH-key based auth won't be triggered on the git clone.  Entering an SSH key passphrase is 1) awkward, and 2) (more importantly) breaks gemini-cli when it tries to `git pull` (I assume) the repo on startup.

Since this is a public repo, there's no need to authenticate to github, and bypassing auth simplifies the flow.